requirements.txt
asgiref==3.5.2 
django==4.1.2 
django-filter==22.1
djangorestframework==3.14.0 
pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3 
tzdata==2022.5